# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Application API
  description: |
    # Agent Application API

    This API provides endpoints for interacting with intelligent agents that implement the ReAct (Reasoning and Acting) pattern. 
    The agents can process natural language requests, reason about tasks, and execute actions using various tools and LLM providers.

    ## Features

    - **Conversational AI**: Create and manage conversations with intelligent agents
    - **ReAct Pattern**: Agents use reasoning and acting cycles to solve complex tasks
    - **Tool Integration**: Agents can use various tools (calculators, web search, file operations)
    - **Multiple LLM Providers**: Support for OpenAI, Anthropic, and Ollama
    - **Real-time Streaming**: Stream agent responses in real-time
    - **Token Management**: Track and limit token usage across conversations
    - **Security**: JWT and API key authentication with role-based access control

    ## Authentication

    The API supports two authentication methods:

    1. **JWT Bearer Token**: Obtain a token via `/auth/login` and include it in the `Authorization` header
    2. **API Key**: Generate an API key via `/auth/api-key` and include it in the `X-API-Key` header

    ## Rate Limiting

    API requests are rate-limited per user and API key. Default limits:
    - 100 requests per minute for authenticated users
    - 1000 requests per hour for API keys
    - Token limits apply per conversation and user

    ## Error Handling
...
servers:
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Authenticates a user with email and password, returning JWT access and refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
            examples:
              login:
                summary: User login
                value:
                  email: user@example.com
                  password: securePassword123
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
              examples:
                success:
                  summary: Successful authentication
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    tokenType: Bearer
                    expiresIn: 86400
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Creates a new user account and returns authentication tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              register:
                summary: User registration
                value:
                  email: newuser@example.com
                  password: securePassword123
                  firstName: John
                  lastName: Doe
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
    AuthenticationRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User's password
          minLength: 8
          example: securePassword123
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: newuser@example.com
        password:
          type: string
          format: password
          description: User's password
          minLength: 8
          example: securePassword123
        firstName:
          type: string
          description: User's first name
          example: John
        lastName:
          type: string
          description: User's last name
          example: Doe
    AuthenticationResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: JWT refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          type: string
          description: Token type
          example: Bearer
        expiresIn:
          type: integer
          description: Token expiration time in seconds
          example: 86400
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: Invalid request data
        details:
          type: array
          description: Detailed error information
          items:
            type: string
          example:
            - Field 'content' is required
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2024-01-01T10:00:00Z'
        path:
          type: string
          description: Request path that caused the error
          example: /conversations/123/messages
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              summary: Validation error
              value:
                error: VALIDATION_ERROR
                message: Invalid request data
                details:
                  - Field 'content' is required
                timestamp: '2024-01-01T10:00:00Z'
                path: /conversations/123/messages
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            auth:
              summary: Authentication required
              value:
                error: UNAUTHORIZED
                message: Authentication required
                timestamp: '2024-01-01T10:00:00Z'
                path: /conversations
...